+++ cmd/mattermost/commands/config.go
@@ -4,7 +4,6 @@
 package commands
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"os"
@@ -15,7 +14,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
-	"github.com/mattermost/mattermost-server/mlog"
 	"github.com/mattermost/mattermost-server/model"
 	"github.com/mattermost/mattermost-server/utils"
 	"github.com/mattermost/mattermost-server/utils/fileutils"
@@ -178,7 +176,7 @@ func configShowCmdF(command *cobra.Command, args []string) error {
 	config := app.Config()
 
 	// pretty print
-	fmt.Printf("%s", prettyPrint(configToMap(*config)))
+	fmt.Printf("%s", prettyPrintStruct(*config))
 	return nil
 }
 
@@ -203,37 +201,6 @@ func printConfigValues(configMap map[string]interface{}, configSetting []string,
 		}
 		return "", fmt.Errorf("%s configuration setting is not in the file", name)
 	}
-}
-
-// prettyPrint the map
-func prettyPrint(configMap map[string]interface{}) string {
-	value := reflect.ValueOf(configMap)
-	return printMap(value, 0)
-}
-
-// printMap takes a reflect.Value and print it out, recursively if its a map with the given tab settings.
-func printMap(value reflect.Value, tabVal int) string {
-
-	out := &bytes.Buffer{}
-
-	for _, key := range value.MapKeys() {
-		val := value.MapIndex(key)
-		if newVal, ok := val.Interface().(map[string]interface{}); !ok {
-			fmt.Fprintf(out, "%s", strings.Repeat("\t", tabVal))
-			fmt.Fprintf(out, "%v: \"%v\"\n", key.Interface(), val.Interface())
-		} else {
-			fmt.Fprintf(out, "%s", strings.Repeat("\t", tabVal))
-			fmt.Fprintf(out, "%v:\n", key.Interface())
-			// going one level in, increase the tab
-			tabVal++
-			fmt.Fprintf(out, "%s", printMap(reflect.ValueOf(newVal), tabVal))
-			// coming back one level, decrease the tab
-			tabVal--
-		}
-	}
-
-	return out.String()
-
 }
 
 func configSetCmdF(command *cobra.Command, args []string) error {
@@ -387,45 +354,4 @@ func UpdateMap(configMap map[string]interface{}, configSettings []string, newVal
 // configToMap converts our config into a map
 func configToMap(s interface{}) map[string]interface{} {
 	return structToMap(s)
-}
-
-// structToMap converts a struct into a map
-func structToMap(t interface{}) map[string]interface{} {
-	defer func() {
-		if r := recover(); r != nil {
-			mlog.Error(fmt.Sprintf("Panicked in structToMap. This should never happen. %v", r))
-		}
-	}()
-
-	val := reflect.ValueOf(t)
-
-	if val.Kind() != reflect.Struct {
-		return nil
-	}
-
-	out := map[string]interface{}{}
-
-	for i := 0; i < val.NumField(); i++ {
-		field := val.Field(i)
-
-		var value interface{}
-
-		switch field.Kind() {
-		case reflect.Struct:
-			value = structToMap(field.Interface())
-		case reflect.Ptr:
-			indirectType := field.Elem()
-
-			if indirectType.Kind() == reflect.Struct {
-				value = structToMap(indirectType.Interface())
-			} else {
-				value = indirectType.Interface()
-			}
-		default:
-			value = field.Interface()
-		}
-
-		out[val.Type().Field(i).Name] = value
-	}
-	return out
 }
